Update to XXH3 for CRC_OPT
