+++ src/CRC32_ARMV8.cpp
@@ -17,34 +17,35 @@ void CRC_Init()
 {
 }
 
-u32 CRC_Calculate( u32 crc, const void * buffer, u32 count )
+u64 CRC_Calculate( u64 crc, const void * buffer, u32 count )
 {
 	u8 *p;
-	u32 orig = crc;
+	u32 crc32 = static_cast<u32>(crc);
+	u32 orig = crc32;
 
 	p = (u8*) buffer;
 
 	// use eight byte crc intrinsic __crc32d if count is high enough.
 	// __crc32d, __crc32w, __crc32h and __crc32b use polynomial 0x04C11DB7
 	while (count >= 8) {
-		crc = __crc32d(crc, *((u64*)p));
+		crc32 = __crc32d(crc32, *((u64*)p));
 		p += 8;
 		count -= 8;
 	}
 	if (count >= 4) {
-		crc = __crc32w(crc, *((u32*)p));
+		crc32 = __crc32w(crc32, *((u32*)p));
 		p += 4;
 		count -= 4;
 	}
 	if (count >= 2) {
-		crc = __crc32h(crc, *((u16*)p));
+		crc32 = __crc32h(crc32, *((u16*)p));
 		p += 2;
 		count -= 2;
 	}
 	if (count == 1)
-		crc = __crc32b(crc, *p);
+		crc32 = __crc32b(crc32, *p);
 
-	return crc ^ orig;
+	return crc32 ^ orig;
 }
 
 u32 CRC_Calculate_Strict( u32 crc, const void * buffer, u32 count )
@@ -52,17 +53,18 @@ u32 CRC_Calculate_Strict( u32 crc, const void * buffer, u32 count )
 	return CRC_Calculate(crc, buffer, count);
 }
 
-u32 CRC_CalculatePalette(u32 crc, const void * buffer, u32 count )
+u64 CRC_CalculatePalette(u64 crc, const void * buffer, u32 count )
 {
 	u8 *p;
-	u32 orig = crc;
+	u32 crc32 = static_cast<u32>(crc);
+	u32 orig = crc32;
 
 	p = (u8*) buffer;
 	while (count--) {
 		// use two byte intrinsic __crc32h
-		crc = __crc32h(crc, *((u16*)p));
+		crc32 = __crc32h(crc32, *((u16*)p));
 		p += 8;
 	}
 
-	return crc ^ orig;
+	return crc32 ^ orig;
 }
